requirements.txt
# Generación de códigos QR
qrcode==7.4.2
pillow==10.4.0

# Generación de PDF
reportlab==4.2.5

# Detección y procesamiento de imágenes
opencv-python==4.10.0.84
numpy==2.1.3

# Lectura de códigos QR/Barras
pyzbar==0.1.9

# Conexión con IP Webcam
requests==2.32.3

# Utilidades
python-json-logger==2.0.7